Add permission judge in RESTful controllers

+++ app/controllers/BoardController.php
@@ -3,6 +3,14 @@
 class BoardController extends \BaseController {
 
 	/**
+	* Instantiate a new UserController instance.
+	*/
+	public function __construct()
+	{
+		$this->beforeFilter('auth', ['only' => ['store', 'update', 'destroy']]);
+	}
+
+	/**
 	 * Display a listing of the resource.
 	 *
 	 * @return Response
@@ -20,6 +28,10 @@ class BoardController extends \BaseController {
 	 */
 	public function store()
 	{
+		if ( !Auth::user()->can('boards_management') ){
+			throw new Exception("Permission Deny", 1);
+		}
+
 		Board::create(array(
 			'type' => Input::get('type'),
 			'code' => Input::get('code'),
@@ -50,11 +62,18 @@ class BoardController extends \BaseController {
 	 */
 	public function update($id)
 	{
+		if ( !Auth::user()->can('boards_management') ){
+			throw new Exception("Permission Deny", 1);
+		}
+
 		$board = Board::find($id)->update(array(
 			'type' => Input::get('type'),
 			'code' => Input::get('code'),
 			'description' => Input::get('description'),
 		));
+
+		// Clear empty elements
+		$update_info = array_diff($update_info, ['']);
 
 		return Response::json(array('success' => true));
 	}
@@ -68,6 +87,10 @@ class BoardController extends \BaseController {
 	 */
 	public function destroy($id)
 	{
+		if ( !Auth::user()->can('boards_management') ){
+			throw new Exception("Permission Deny", 1);
+		}
+
 		Board::destroy($id);
 		return Response::json(array('success' => true));
 	}
