+++ app/controllers/UserController.php
@@ -3,6 +3,14 @@
 class UserController extends \BaseController {
 
 	/**
+	* Instantiate a new UserController instance.
+	*/
+	public function __construct()
+	{
+		$this->beforeFilter('auth', ['only' => ['store', 'update', 'destroy']]);
+	}
+
+	/**
 	 * Display a listing of the resource.
 	 *
 	 * @return Response
@@ -20,6 +28,12 @@ class UserController extends \BaseController {
 	 */
 	public function store()
 	{
+		if ( !Auth::user()->can('users_management') ){
+			throw new Exception("Permission Deny", 1);
+		}
+
+		// TODO: Form Validation
+
 		User::create(array(
 			'username' 		=> Input::get('username'),
 			'password' 		=> Input::get('password'),
@@ -52,14 +66,30 @@ class UserController extends \BaseController {
 	 */
 	public function update($id)
 	{
-		User::find($id)->updatte(array(
-			'username' 		=> Input::get('username'),
+		if ( !( Auth::user()->can('users_management') OR ($id !== Auth::id()) ) ){
+			throw new Exception("Permission Deny", 1);
+		}
+
+		// TODO: Form Validation
+
+		$update_info = array(
 			'password' 		=> Input::get('password'),
 			'title' 		=> Input::get('title'),
-			'roles' 		=> Input::get('roles'),
 			'email' 		=> Input::get('email'),
 			'phone' 		=> Input::get('phone'),
-		));
+		);
+
+		if ( Auth::user()->can('users_management') ) {
+			$update_info = array_merge($update_info, [
+				'username' 		=> Input::get('username'),
+				'phone' 		=> Input::get('phone'),
+			]);
+		}
+
+		// Clear empty elements
+		$update_info = array_diff($update_info, ['']);
+
+		User::find($id)->update($update_info);
 
 		return Response::json(array('success' => true));
 	}
@@ -73,9 +103,12 @@ class UserController extends \BaseController {
 	 */
 	public function destroy($id)
 	{
+		if ( !( Auth::user()->can('users_management') OR ($id !== Auth::id()) ) ){
+			throw new Exception("Permission Deny", 1);
+		}
+
 		$user = User::destroy($id);
         return Response::json(array('success' => true));
 	}
-
 
 }
